Silence warning.
git-svn-id: https://src.develer.com/svnoss/bertos/trunk@3984 38d2e660-2303-0410-9eaa-f027e97ec537

+++ bertos/cfg/kfile_debug.c
@@ -46,7 +46,7 @@ static size_t kfiledebug_write(struct KFile *_fd, const void *buf, size_t size)
 {
 	KFILEDEBUG_CAST(_fd);
 
-	kprintf("%.*s", size, (const char *)buf);
+	kprintf("%.*s", (int)size, (const char *)buf);
 
 	return size;
 }
